Move db package into service/
NB some not very obvious places where paths must be changed, notably
the tests (which refer to migrations at a particular relative path)
and the Makefile (which refers to migrations at a particular relative
path).
